Use th:each to show a list of DTOs in a table.
Watch this video to see how this commit was implemented: https://youtu.be/1kouIWceI-A

Demonstrate how to show a list of objects from Thymeleaf into a Boostrap list-group.  We take a list of objects and make them available to our HTML page using th:each.  Then, we add a hyperlink to the table, using th:href and the Thymeleaf hyperlink syntax.  We also show how to represent an object by overriding the toString() method of java.lang.Object
Next, we create dynamic URL by using the th:href attribute, and a bit of syntax to pull a value out of an object, and place it in a URL.